+++ internal/ecosystem/packagist/const.go
@@ -1,0 +1,7 @@
+package packagist
+
+const (
+	EcosystemName    = "Packagist"
+	ComposerFile     = "composer.json"
+	ComposerLockFile = "composer.lock"
+)