+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.0
 
 - Add go package detection
+- Add packagist package detection
 - Auto-Update database if needed when using scan subcommand
 
 # 0.1.0